Massive refactor and some tweaks
