+++ src/geocoder/azure.rs
@@ -1,0 +1,72 @@
+use std::env;
+use reqwest::Client;
+use serde::Deserialize;
+use crate::GenericError;
+use crate::geocoder::geoposition::{GeoPosition, GeocoderRequest};
+
+#[derive(Deserialize, Copy, Clone, Debug)]
+pub struct AzureGeocoderPosition {
+    pub lat: f64,
+    pub lon: f64,
+}
+
+#[derive(Deserialize)]
+struct AzureGeocoderResult {
+    id: String,
+    r#type: String,
+    position: AzureGeocoderPosition,
+}
+
+#[derive(Deserialize)]
+struct AzureGeocoderResponse {
+    results: Vec<AzureGeocoderResult>,
+}
+
+pub trait AzureGeocoderRequest {
+    async fn azure_geocoder_request(&self) -> Result<AzureGeocoderPosition, GenericError>;
+}
+
+impl From<AzureGeocoderResponse> for Result<AzureGeocoderPosition, GenericError> {
+    fn from(value: AzureGeocoderResponse) -> Self {
+        // We are *likely* to want a cross-street value
+        for result in &value.results {
+            if result.r#type == "Cross Street" {
+                return Ok(result.position);
+            }
+        }
+
+        // If there is none, the first result is all we've got
+        if let Some(result) = value.results.first() {
+            Ok(result.position)
+        } else {
+            Err("Geocoder returned no results".into())
+        }
+    }
+}
+
+impl AzureGeocoderRequest for GeocoderRequest {
+    async fn azure_geocoder_request(&self) -> Result<AzureGeocoderPosition, GenericError> {
+        if let Ok(api_key) = env::var("AZURE_API_KEY") {
+            let client = Client::new();
+            let req = format!("{}, VICTORIA, AUSTRALIA", self.request);
+            let request = format!("https://atlas.microsoft.com/search/address/json?&subscription-key={api_key}&api-version=1.0&language=en-AU&countrySet=AU&query={req}");
+            let res = client
+                .get(request)
+                .send()
+                .await?;
+
+            let body = res.json::<AzureGeocoderResponse>().await?;
+            return body.into();
+        }
+        Err("Geocoding failed".into())
+    }
+}
+
+impl From<&AzureGeocoderPosition> for GeoPosition {
+    fn from(value: &AzureGeocoderPosition) -> Self {
+        Self {
+            latitude: value.lat,
+            longitude: value.lon,
+        }
+    }
+}
