+++ src/geocoder/mod.rs
@@ -1,0 +1,3 @@
+pub mod geoposition;
+pub mod google;
+pub mod azure;
